+++ database/migrations/2019_07_06_121906_add_category_id_column_to_items_table.php
@@ -14,7 +14,7 @@ class AddCategoryIdColumnToItemsTable extends Migration
     public function up()
     {
         Schema::table('items', function (Blueprint $table) {
-          $table->integer('category_id')->unsigned()->nullable();
+          $table->bigInteger('category_id')->unsigned()->nullable();
           $table->foreign('category_id')->references('id')->on('categories');
         });
     }
